refactor(project): refactor all the project in order to have version controls and clean dev env
