+++ README.md
@@ -1,4 +1,26 @@
-# Item Spreadsheet to JSON
+<h1 align="center">
+	<img src="https://i.imgur.com/09kSsRP.png" />
+</h1>
+<p align="center">  
+<a href="https://david-dm.org/@nooksbazaar/acdb">  
+    <img alt="" src="https://david-dm.org/@nooksbazaar/acdb/status.svg?style=flat-square">  
+</a>  
+<a href="https://prettier.io/">  
+    <img alt="" src="https://img.shields.io/badge/code_style-prettier-ff69b4.svg?style=flat-square">  
+</a> 
+<br>  
+<a href="https://npmjs.com/package/@stun3r/acnh-translations">  
+    <img alt="" src="https://img.shields.io/npm/v/@nooksbazaar/acdb/latest.svg?style=flat-square">  
+</a>  
+<a href="https://npmjs.com/package/@stun3r/acnh-translations">  
+    <img alt="" src="https://img.shields.io/npm/dt/@nooksbazaar/acdb.svg?style=flat-square">  
+</a>  
+</p>  
+  
+Our goal is to provide to the Animal Crossing community the best set of tools to create beautiful API and project around the Animal Crossing: New Horizons [community spreadsheet](https://docs.google.com/spreadsheets/d/13d_LAJPlxMa_DubPTuirkIV4DERBMXbrWQsmSh8ReK4/edit)) into JSON
 
-This repository contains the code used to convert the Animal Crossing: New Horizons community items
-google [spreadsheet](https://docs.google.com/spreadsheets/d/13d_LAJPlxMa_DubPTuirkIV4DERBMXbrWQsmSh8ReK4/edit?usp=sharing) into JSON.
+## License
+
+[MIT License](./LICENSE) - Copyright (c) Adam Lavin & Thibaut DAVID
+
+> ACDB is a fan-made project and is in no way affiliated with Nintendo.
