--- bin/publish.sh
@@ -1,7 +1,0 @@
-#!/usr/bin/env bash
-
-cp package.json out
-
-npm publish out --access=public
-
-rm out/package.json
