+++ scripts/publish.sh
@@ -1,0 +1,7 @@
+#!/usr/bin/env bash
+
+cp package.json out
+
+npm publish out --access=public
+
+rm out/package.json
